% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{stocc-package}
\alias{stocc}
\alias{stocc-package}
\title{R package for fitting spatial-temporal occupancy models via Gibbs sampling}
\description{
This package contains functions that fit a spatial occupancy model where the
true occupancy is a function of a spatial process.  An efficient Gibbs
sampling algorithm is used by formulating the detection and occupancy
process models with a probit model instead of the traditional logit based
model.
}
\details{
\tabular{ll}{ Package: \tab stocc\cr Type: \tab Package\cr Version: \tab
1.23\cr Date: \tab 7-29-2014\cr License: \tab Unlimited\cr LazyLoad: \tab
yes\cr }
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}

