% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stocc-package.R
\docType{package}
\name{stocc-package}
\alias{stocc}
\alias{stocc-package}
\title{Fit a Spatial Occupancy Model via Gibbs Sampling}
\description{
This package contains functions that fit a spatial occupancy model where the
true occupancy is a function of a spatial process.  An efficient Gibbs
sampling algorithm is used by formulating the detection and occupancy
process models with a probit model instead of the traditional logit based
model.
}
\details{
\tabular{ll}{ Package: \tab stocc\cr Type: \tab Package\cr Version: \tab
1.30\cr Date: \tab August 21, 2015\cr License: \tab file LICENSE\cr LazyLoad: \tab
yes\cr }
}
\author{
Devin S. Johnson

Maintainer: Devin S. Johnson <devin.johnson@noaa.gov>
}

